requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.0
faiss-cpu>=1.8.0
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
openai>=1.0.0
huggingface-hub>=0.20.0
cohere>=5.0.0
groq>=0.4.0

